pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["Jatin Roopchandani <jatinroopchandani3007@gmail.com>"]
readme = "README.md"
packages = [{include = "flask-app"}] 
[tool.poetry.dependencies]
python = ">=3.12,<4.0"
python-dotenv = ">=1.0.1,<2.0.0"
pyjwt = ">=2.10.1,<3.0.0"
supabase = ">=2.13.0,<3.0.0"
flask = "^3.1.0"
google = "^3.0.0"
google-genai = "^1.4.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
